implement improvements as commented on github
