+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinLanguageManager.java
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.mixins.transformers;
 
-import at.hannibal2.skyhanni.events.minecraft.LanguageChangeEvent;
+import at.hannibal2.skyhanni.events.minecraft.MinecraftLanguageChangeEvent;
 import net.minecraft.client.resources.Language;
 import net.minecraft.client.resources.LanguageManager;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,9 +9,9 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(LanguageManager.class)
-public class LanguageChangeEventPatch {
+public class MixinLanguageManager {
     @Inject(method = "setCurrentLanguage", at = @At("TAIL"))
     private void onLanguageSet(Language currentLanguageIn, CallbackInfo ci) {
-        new LanguageChangeEvent(currentLanguageIn).post();
+        new MinecraftLanguageChangeEvent(currentLanguageIn).post();
     }
 }
